[RESTRICT AUTOMERGE] TJBench: Fix errors when decomp. files w/ ICC data
This is a backport of an upstream commit. Original commit message
included the following:

"""
Embedded ICC profiles can cause the size of a JPEG file to exceed the
size returned by tjBufSize() (which is really meant to be used for
compression anyhow, not for decompression), and this was causing a
segfault (C) or an ArrayIndexOutOfBoundsException (Java) when
decompressing such files with TJBench.  This commit modifies the
benchmark such that, when tiled decompression is disabled, it re-uses
the source buffer as the primary JPEG buffer.
"""

This commit was pulled in due to a SEGFAULT that appeared after
backporting the fix to the original bug.

Bug: 120551338
Test: tj64 /data/local/tmp/crash-46.jpg
See b/120551338#comment1 to get crash-46.jpg

Change-Id: Ia480d23a5581027da11a2da168c06448161668bf
(cherry picked from commit 87ccce3a16a600ff52b91741421f23583bc146e3)